backend/requirements.txt
fastapi
uvicorn
websockets
google-cloud-speech
google-genai
asyncpg
python-jose[cryptography]
python-dotenv
httpx
sarvamai
 